PILOT1_RDN/ai4eu/requirements.txt
grpcio==1.41.1
grpcio-tools==1.41.1
protobuf==3.17.3
PyYAML==5.4.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
numpy==1.19.5
tqdm==4.62.2
python-dotenv==0.19.2
u8darts[torch]==0.11.0